Refactor the server
